Use GLfloat instead of C float for everything GL-related

+++ genmaze.c
@@ -158,7 +158,7 @@ VERTICES_SET get_vertices(const int size, int *lines[])
 			if (lines[i][j] == 0)
 			{
 				int u = i, v = j;
-				VERTEX vtx = {.x = j, .y = -0.5f, .z = i};
+				VERTEX vtx = {.x = (GLfloat)j, .y = (GLfloat)-0.5, .z = (GLfloat)i};
 				int direction = 2;
 				GLsizei v_index_begin = v_index;
 
@@ -175,15 +175,15 @@ VERTICES_SET get_vertices(const int size, int *lines[])
 					}
 
 					// Makes (x, z)==(0, 0) the center of the maze
-					vtx.x-= size / 2.0f;
-					vtx.z-= size / 2.0f;
+					vtx.x-= size / (GLfloat)2.0;
+					vtx.z-= size / (GLfloat)2.0;
 
 					// Starts to define the first triangle
 					*i_ptr++ = v_index;
-					vertices[v_index++] = vtx;    // stores the "key" vertex (y == -0.5)
+					vertices[v_index++] = vtx;            // stores the "key" vertex (y == -0.5)
 					*i_ptr++ = v_index;
 					vertices[v_index] = vtx;
-					vertices[v_index++].y = 0.5f; // stores the corresponding vertex (y == 0.5)
+					vertices[v_index++].y = (GLfloat)0.5; // stores the corresponding vertex (y == 0.5)
 
 					switch (direction)
 					{
